requirements/base.txt
-c constraints.txt
Django==1.11.16
wagtail==1.13.1

psycopg2==2.7.5

django-compressor==2.2
wagtail-django-recaptcha==1.0
wagtailgmaps==0.4
slackweb==1.0.5
requests[security]
django-allauth

raven
